Remove incomplete features to import data
This feature was very useful for me as beta testeur.

Accomodations, registries and services were listed in specific sheet,
 with custom format. I just had to save these files as JSON, and import
 these data as seeds.

It was very useful, and allow us to save a loooot of time...

But it would be a lot of work to develop this service for classic users.
I'd glad to develop this, it would be a great improvement :)